+++ pkg/providers/mtlstruststore/mtlstruststore.go
@@ -1,0 +1,7 @@
+package mtlstruststore
+
+import "github.com/akamai/terraform-provider-akamai/v7/pkg/providers/registry"
+
+func init() {
+	registry.RegisterSubprovider(NewSubprovider())
+}